+++ src/JuES.jl
@@ -8,8 +8,8 @@ export psi4
 include("Input/Input.jl")
 include("DiskTensors/DiskTensors.jl")
 include("Backend/Transformation.jl")
-include("Backend/Integrals.jl")
 include("Backend/Wavefunction.jl")
+include("Backend/IntegralTransformation.jl")
 include("Backend/DF.jl")
 include("Backend/Direct.jl")
 include("ConfigurationInteraction/Determinant.jl")
